添加后台支付逻辑

+++ 01.项目源码/trunk/0.0.1/clothing-server/src/main/java/com/dower/sharerideapp/controller/weichat/WXPayController.java
@@ -112,7 +112,7 @@ public class WXPayController {
                 ClProduct cloth = (ClProduct) clothing.getResultInfo();
                 String vcExpireTime = cloth.getVcExpireTime();
                 String vcOrderNo = cloth.getVcOrderNo();
-                BigDecimal numPrice = cloth.getNumPrice();
+                Long numPrice = cloth.getNumPrice();
 
                 MyConfig config = new MyConfig();
                 WXPay wxpay = new WXPay(config);
